Fixes in Vox loader, Impoved physics perfomance

+++ VoxCake/Common/Voxel.cs
@@ -2,15 +2,15 @@
 {
     public struct Voxel
     {
-        public short x, y, z;
+        public int x, y, z;
         public uint value;
         public bool processed;
         
         public Voxel(int x, int y, int z, uint value, bool processed)
         {
-            this.x = (short)x;
-            this.y = (short)y;
-            this.z = (short)z;
+            this.x = x;
+            this.y = y;
+            this.z = z;
             this.value = value;
             this.processed = processed;
         }
